Major Restructure
moisture_rnn now custom class, old stuff for reproducibility moved to version_control
